Cleaned Config.xml to be more convienient for git.
The Version and LastModified tags are empty. They
will be populated when versions are published, but
versions should never be committed to the repo.

The Authors tag simply contains a reference to the
GitHub page. No addition/deletion notes will be
made in this file; it is assumed that anyone who
cares will say `git log` (or the GitHub equivalent).